requirements.txt
# Flask and Core Dependencies
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-migrate>=4.0.0
flask-cors>=4.0.0

# Authentication
msal>=1.24.0

# Database
sqlalchemy>=2.0.0
pyodbc>=4.0.39

# Data Processing
pandas>=2.0.0
openpyxl>=3.1.0
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
werkzeug>=3.0.0

# Development
pytest>=7.4.0
pytest-flask>=1.3.0
black>=23.7.0
flake8>=6.0.0

# Production
gunicorn>=21.2.0